GB-83 fix: 마이페이지 서비스 로직 분리
